backend/requirements.txt
flask
flask_cors
gunicorn
nltk
scikit-learn
numpy
gensim
pytesseract
pillow
googletrans==4.0.0-rc1
python-dotenv